Modularize concurrent calls logic
There is a module info/concurrent_calls.mo which contains all the logic of making concurrent calls, buffering the futures, awaiting the futures, catching errors, and detecting a trap in a continuation with a finally clause.

Because async generics do not exist in Motoko this module can not be a generic package. To make it as re-usable as possible it imports another module, info/single_call.mo, which exports the shared function that is the single call. It then wraps all the concurrency logic around the imported single calls.

The client code passes a list of call args in for all concurrent calls and a list of continuation callbacks, separated for the success case and the async error case.

There are also callbacks for each time a concurrent call is scheduled. Because it can happen that somewhere in the middle scheduling starts to fail because of a full queue. Through the callbacks the client code knows which ones have been scheduled.

The is also a single callback (not one for each concurrent call) for when a trap in a continuation is caught. It can normally only be used to shut down operations. As a rule, the continuation callbacks passed in by the client must not trap. If one of the callbacks traps then all subsequent futures will not be awaited anymore. They will be lost and the client code will not learn about there results. It will not even learn when or if they returned. The trap will be detected though and the trap callback called with a Nat value which is the number of calls that have been awaited so far and whose continuation did not trap.

The CanisterHistory class and main.mo code are adjusted to use the module.