Simplify sync scheduling without Deque
The Deque is more precise but we remove it here for a simpler
logic which just goes round robin over all indices of the history
storage.

For example, newly added canisters might be scheduled for their
second time earlier than before, etc. But we save the additional
data structure.

TODO: if a sync fails due to an async error in the call to the
mgmt canister (canister_info) then with current logic it would
get skipped for this round. This can be improved by putting it
into a "backlog" list. During scheduling the backlog list (if
non-empty) would be processed first before continuing with the
round robin logic. But before doing this we have to understand
all possible async errors better and in detail. We don't want to
add something that would permanently fail (such as a deleted
canister) because then prioritizing the backlog would hold up
progress forever.